Fix type and formatting errors found by Dartium
R=rchandia@google.com
BUG=
TEST=

Review URL: https://chromereviews.googleplex.com/3543021

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@152 260f80e4-7a28-3924-810f-c04153c831b5

+++ dart/client/samples/total/src/SelectionManager.dart
@@ -150,10 +150,10 @@ class SelectionManager {
     ClientRect minP = minCellElmt.getBoundingClientRect();
     ClientRect maxP = maxCellElmt.getBoundingClientRect();
     return new BoundingBox(
-        minP.left - orgP.left,
-        minP.top - orgP.top,
-        maxP.left - minP.left + maxCellElmt.clientWidth,
-        maxP.top - minP.top + maxCellElmt.clientHeight);
+        (minP.left - orgP.left).toInt(),
+        (minP.top - orgP.top).toInt(),
+        (maxP.left - minP.left + maxCellElmt.clientWidth).toInt(),
+        (maxP.top - minP.top + maxCellElmt.clientHeight).toInt());
   }
 
   CellRange getSelectionRange() {
